// utils/enum.ts
export enum ETagType {
  h1 = "H1",
  h2 = "H2",
  h3 = "H3",
  h4 = "H4",
  h5 = "H5",
  h6 = "H6",
  p = "P",
  span = "SPAN",
  div = "DIV",
  img = "IMG",
  a = "A",
}

export enum TITLE {
  CONTAINER = "Container Setting",
  HEADER = "Header Setting",
  CONTENT = "Content Setting",
  IMAGE = 'Image Setting'
}

export enum TITLE_COLOR {
  COLOR = "Color",
  BACKGROUND_COLOR = "BackgroundColor",
}